ui/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
openai==1.56.2
